+++ bold-pagos-en-linea/src/index.js
@@ -1,1 +1,2 @@
 import './components/ButtonBlock';
+import './utils/libraries/global.sweetalert2'